backend/requirements.txt
fastapi==0.129.0
uvicorn[standard]==0.34.0
pydantic==2.10.5
pydantic-settings==2.13.0
python-multipart==0.0.22
PyPDF2==3.0.1
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==11.0.0
python-dotenv==1.0.1
